Renaming envitest to org.enso.os.environment.lib
