+++ lib/java/os-environment-lib/src/main/java/org/enso/os/environment/lib/HelloTitle.java
@@ -1,4 +1,4 @@
-package org.enso.os.envitest;
+package org.enso.os.environment.lib;
 
 import java.io.FileWriter;
 import java.util.function.Function;
@@ -7,7 +7,7 @@ import org.enso.persist.Persistable;
 import org.enso.persist.Persistance;
 
 /** Entry point for a "isolate library" to be loaded and communicated to via a {@link Channel}. */
-public final class EnviMain extends Channel.Config {
+public final class HelloTitle extends Channel.Config {
   public static void main(String... args) throws Exception {
     try (java.io.FileWriter out = new FileWriter(args[0])) {
       out.write(args[1]);
@@ -21,9 +21,9 @@ public final class EnviMain extends Channel.Config {
 
   /** This is a message that os-environment/test sends from to dynamic library. */
   @Persistable(id = 4332)
-  public static record Hello(String name) implements Function<Channel<EnviMain>, Text> {
+  public static record Hello(String name) implements Function<Channel<HelloTitle>, Text> {
     @Override
-    public Text apply(Channel<EnviMain> ch) {
+    public Text apply(Channel<HelloTitle> ch) {
       // now we are running in library and sending a message back to executable
       var withTitle = ch.execute(Text.class, new Title(name));
       return new Text("Hello " + withTitle.msg() + "!");
