+++ lib/java/os-environment-lib/src/main/resources/META-INF/native-image/jni-config.json
@@ -1,6 +1,6 @@
 [
   {
-    "name": "org.enso.os.envitest.EnviMain",
+    "name": "org.enso.os.environment.lib.HelloTitle",
     "methods": [
       {
         "name": "main",
